🔨 : 불필요 tsx 삭제 및 RealBoard -> Board로 수정
